fix: address PR feedback and improve code quality
High priority fixes:
- Fix DualQueryLogger cache hit logging in llm_chain.py to use correct method signature
- Pass cache_hit=True in metadata for proper SQLite logging to admin dashboard

Code quality improvements:
- Remove duplicate IndexedDocument and IndexedDocumentsResponse models from content.py
- Import shared models from knowledge.py to follow DRY principle
- Add exception chaining with 'from e' for better error diagnosis
- Improve debugging capabilities in content.py error handlers

All cached responses now properly logged to both JSON and SQLite databases.
Exception tracebacks preserved for easier troubleshooting.

+++ backend/core/query_data_manager.py
@@ -115,7 +115,7 @@ class QueryDataManager:
             logger.error(f"Error initializing query data manager database: {str(e)}", exc_info=True)
             raise
 
-    def get_overview_stats(self, days: int = 7) -> Dict[str, Any]:
+    def get_overview_stats(self, days: float = 7) -> Dict[str, Any]:
         """Get overview statistics for the specified number of days, including comparison with previous period."""
         try:
             with self.get_connection() as conn:
